+++ ErsatzTV.Core/Domain/GraphicsElement.cs
@@ -1,3 +1,5 @@
+using ErsatzTV.Core.Domain.Scheduling;
+
 namespace ErsatzTV.Core.Domain;
 
 public class GraphicsElement
@@ -9,4 +11,9 @@ public class GraphicsElement
     public List<PlayoutItemGraphicsElement> PlayoutItemGraphicsElements { get; set; }
     public List<ProgramScheduleItem> ProgramScheduleItems { get; set; }
     public List<ProgramScheduleItemGraphicsElement> ProgramScheduleItemGraphicsElements { get; set; }
+    public List<Deco> Decos { get; set; }
+    public List<DecoGraphicsElement> DecoGraphicsElements { get; set; }
+
+    // for unit testing
+    public override string ToString() => Path;
 }